requirements.txt
certifi==2025.1.31
dnspython==2.7.0
pymongo==4.3.3
python-dotenv==1.1.0
requests==2.32.3
